refactor v1 into separate module
